Downloader with command line argument
